README.md: deduplicate example code
This commit adds a Makefile target for generating the README, which
allows us to always keep the sample code and the embedded sample code in
sync. Without this commit, changes to the example, like those made in
the previous commit, would put the example out of sync with the embedded
code.

The new makefile target will replace all code samples preceeded by
`[replace] # (<file>)` with the literal contents of `<file>`.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>