Dockerfile: remove run-server.sh
Now that meowlflow is more than just a sidecar, it does not make sense
for the entrypoint to the meowlflow container to be the run-server.sh
script. Furthermore, that script does not allow passing all of the
necessary parameters to the sidecare sub-command, which means that the
script must always be bypassed and is this not really useful. This
commit removes the run-server.sh script entirely. Instead, the
entrypoint for the container is set to meowlflow, enabling more
idiomatic usage of the container, e.g.
`docker run --rm img.conny.dev/conny/meowlflow generate ...`.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>