meowflow/build.py: accept --workdir flag
Currently, `meowlflow generate` uses the mlflow TempDir util, which
creates a temporary directory upon entering the block and destroys the
temporary directory when leaving the block. This does not make sense for
`meowlflow generate` as the generated Dockerfile relies on the model
being available when the Docker container is actually built. If the
temporary directory is removed, then the container cannot be built. This
commit fixes this issue by adding an optional `--workdir` flag into
which the model will be downloaded. The flag defaults to the current
working directory.

Furthermore, this commit ensures that the absolute path to the model is
used in the generated Dockerfile. `meowlflow build` currently works
because the `exec` call switches to the temporary directory and uses
this as its working directory, ensuring that `COPY model_dir...` works
with relative paths. This is not possible if there is no `exec` call,
i.e. in `meowlflow generate`.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>