meowlflow/build.py: use relative path Dockerfile
Container building forces that the only files available to the build are
files that are within the current directory. All absolute paths used
during build are translated into paths relative to the current
directory.
This means that the directive:
```dockerfile
COPY /x/y .
```
will actually copy the file `$PWD/x/y` into the container.

Because of this, we have to use _only_ `model_dir` as the directory we
want to copy into the container. This will work with `meowlflow build`
and `meowlflow generate && docker build .` provided that the `build`
step is executed from the directory specified with the `--workdir` flag.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>