meowlflow: fix OpenAPI operation ID
Since the method that is decorated with the FastAPI router was renamed
to `_infer`, the auto-generated OpenAPI summary and operation ID were
renamed to ` Infer` and `_infer_api_v1_infer_post` respectively, from
the nicer `Infer` and `infer_api_v1_infer_post`. This commit changes the
method name back to `infer` so that the generated OpenAPI specification
includes more aesthetic values.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>