meowflow: allow customizing model path
Currently, when loading the model to serve via `meowlflow serve`, the
model path is always assumed to be `/opt/mlflow/model`, however, it's
possible that the model is located elsewhere, e.g. during local
development. This commit adds a new flag to the `serve` command to allow
specifying the model path.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>