meowlflow: fix OpenAPI generation
Currently, the OpenAPI specifications generated by Meowlflow ignore the
title, description, version, etc that are configured in the schema. This
is a bug due to setting the attributes on the wrong object. This commit
fixes the generation of OpenAPI specs by setting the discovered
attributes on the `app` object, which allows FastAPI's automatic
generation to work correctly.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>