Dockerfile: clean up
This commit fixes the way that we install the local package. Previsouly,
`pip install .` was executed _after_ all of the build-tooling was
uninstalled. This did not fail because we conveniently have no
Git dependencies int our setup.cfg. If we did happen to be tracking
some Git dependencies, this step would fail because Git was no longer
installed. This commit fixes the bug by reordering the `pip install .`
step before the removal of these build tools.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>