e2e: add test for OpenAPI
This commit refactors the e2e test suite and adds a test to validate the
generation of Open API specifications. It also enhances the CI workflow
with a check to ensure that the schema in the examples directory is
up-to-date.

With this refactored e2e test suite we can more easily test scenarios
like `meowlflow sidecar` or `meowlflow promote` and benefit from the
existing logic that automatically spins up an MLflow server before every
test and later cleans it up.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>