Clean Up Repository and Add Types (#46)
* meowlflow: remove unused module

The loghandler module is not used as this package uses the standard
library logging package.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* dependencies: add mypy

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* *: add types

This commit finishes adding type hints to all of the modules in
meowlflow.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* README: regenerate

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* Makefile: simplify Makefile

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* e2e: add e2e test

This commit adds a complete end-to-end test for `meowlflow serve`. The
test will:
1. start mlflow;
2. create an experiment run using the mlflow-example repository;
3. serve the model using meowlflow; and
4. verify that the prediction from the model matches our expectation.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

* e2e: add test for OpenAPI

This commit refactors the e2e test suite and adds a test to validate the
generation of Open API specifications. It also enhances the CI workflow
with a check to ensure that the schema in the examples directory is
up-to-date.

With this refactored e2e test suite we can more easily test scenarios
like `meowlflow sidecar` or `meowlflow promote` and benefit from the
existing logic that automatically spins up an MLflow server before every
test and later cleans it up.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>