e2e: add scenario for generate
This commit enhances the e2e test suite with a new test for the
`meowlflow generate` command. In writing the tests I realized that it
would be an easy win to add a new option to the `build` and `generate`
commands to optionally copy the model schema into the container.
Currently, whenever we run a meowlflow container we have to mount a
volume with the schema, which works totally fine, but is potentially an
additional source of complexity. When the `--schema-path` option is
provided, the Dockerfile is modified to include extra directives to copy
the schema to the cannonical meowlflow schema path, i.e.
`/var/lib/meowlflow/schema.py`. This allows baking a versioned schema
into a container alongside a versioned model, which is nice for
reproducibility. The option of mounting the schema into the container at
run-time is of course still available.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>