meowlflow: remove import for side effects
Previously, meowlflow used pipreqs to generate the list of used
packages in a requirements.txt without collecting everything from the
environment. In order to do this, pipreqs analyzes all of the import
statements in the package. Because the project depended on boto3 but it
was not explicitly used, an import needed to be added for the side
effect of indicating to pipreqs that this package was indeed a
dependency. However, now that this package uses Poetry for its
dependency management we can explicitly add boto3 to the dependency list
and remove the unused import.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>