CI: add tooling to distribute project
This commit adds tooling to CI to automatically build wheels for the
project and publish them to pypi. This is done using the
`pypa/gh-action-pypi-publish` GitHub action. Furthermore, versioning for
the project is simplified to _always_ use Git tags as the source of
truth. Previsouly, the version in pyproject.toml and under
meowlflow/__init__.py had to be manually kept up to date; this is now
done automatically using the `poetry-dynamic-versioning` poetry plugin.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>