meowlflow/promote.py: automatically create models
This commit enhances meowlflow so that models are automatically created
during promotion if they do not yet exist. This matches the expected
user experience of allowing promotion when a model exists but no version
has been staged yet, which essentially asserts that the current run must
be better than nothing.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>